Make BGPError operate on Uint8
Error codes are unsigned bytes, use the proper encapsulation.

+++ bgp/rib-impl/src/test/java/org/opendaylight/protocol/bgp/rib/impl/BGPSessionImplTest.java
@@ -181,8 +181,8 @@ public class BGPSessionImplTest {
         assertEquals(1, this.receivedMsgs.size());
         assertTrue(this.receivedMsgs.get(0) instanceof Notify);
         final Notify error = (Notify) this.receivedMsgs.get(0);
-        assertEquals(BGPError.CEASE.getCode(), error.getErrorCode().shortValue());
-        assertEquals(BGPError.CEASE.getSubcode(), error.getErrorSubcode().shortValue());
+        assertEquals(BGPError.CEASE.getCode(), error.getErrorCode());
+        assertEquals(BGPError.CEASE.getSubcode(), error.getErrorSubcode());
         verify(this.speakerListener).close();
     }
 
@@ -193,8 +193,8 @@ public class BGPSessionImplTest {
         assertEquals(1, this.receivedMsgs.size());
         assertTrue(this.receivedMsgs.get(0) instanceof Notify);
         final Notify error = (Notify) this.receivedMsgs.get(0);
-        assertEquals(BGPError.FSM_ERROR.getCode(), error.getErrorCode().shortValue());
-        assertEquals(BGPError.FSM_ERROR.getSubcode(), error.getErrorSubcode().shortValue());
+        assertEquals(BGPError.FSM_ERROR.getCode(), error.getErrorCode());
+        assertEquals(BGPError.FSM_ERROR.getSubcode(), error.getErrorSubcode());
         verify(this.speakerListener).close();
     }
 
@@ -221,8 +221,8 @@ public class BGPSessionImplTest {
         assertEquals(3, this.receivedMsgs.size());
         assertTrue(this.receivedMsgs.get(2) instanceof Notify);
         final Notify error = (Notify) this.receivedMsgs.get(2);
-        assertEquals(BGPError.HOLD_TIMER_EXPIRED.getCode(), error.getErrorCode().shortValue());
-        assertEquals(BGPError.HOLD_TIMER_EXPIRED.getSubcode(), error.getErrorSubcode().shortValue());
+        assertEquals(BGPError.HOLD_TIMER_EXPIRED.getCode(), error.getErrorCode());
+        assertEquals(BGPError.HOLD_TIMER_EXPIRED.getSubcode(), error.getErrorSubcode());
         verify(this.speakerListener).close();
     }
